refactored and separated execute_utils
